==> src/01-extract-embed-ner.ipynb <==
import pandas as pd
import numpy as np
from gensim.models import Word2Vec, FastText
import spacy

# MEDICAL NER EXTRACTION USING MED7

med7 = spacy.load("en_core_med7_lg")
new_notes = pd.read_pickle("../data/clinical_notes/preprocessed_notes.p")
new_notes['ner'] = None

count = 0
preprocessed_index = {}
patient_clinical_notes = {}
for i in new_notes.itertuples():

    if count % 1000 == 0:
        print(count)

    count += 1
    ind = i.Index
    text = i.preprocessed_text

    if p_id not in patient_clinical_notes:
        patient_clinical_notes[p_id] = []

    all_pred = []
    for each_sent in text:
        try:
            doc = med7(each_sent)
            result = ([(ent.text, ent.label_) for ent in doc.ents])
            result_only = [ent.text for ent in doc.ents]
            if len(result) == 0: continue
            all_pred.append(result)
            patient_clinical_notes[p_id].extend(result)
        except:
            print("error..")
            continue
    new_notes.at[ind, 'ner'] = all_pred

pd.to_pickle(new_notes, "../data/clinical_notes/ner_df.p")
pd.to_pickle(patient_clinical_notes,"../data/clinical_notes/patient_clinical_notes.pkl")

def mean(a):
    return sum(a) / len(a)

##Patient ID -> NER words dictionary

null_index_list = []
for i in new_notes.itertuples():

    if len(i.ner) == 0:
        null_index_list.append(i.Index)
new_notes.drop(null_index_list, inplace=True)

med7_ner_data = {}

for ii in new_notes.itertuples():

    p_id = ii.SUBJECT_ID
    ind = ii.Index

    try:
        new_ner = new_notes.loc[ind].ner
    except:
        new_ner = []

    unique = set()
    new_temp = []

    for j in new_ner:
        for k in j:
            unique.add(k[0])
            new_temp.append(k)

    if p_id in med7_ner_data:
        for i in new_temp:
            med7_ner_data[p_id].append(i)
    else:
        med7_ner_data[p_id] = new_temp

pd.to_pickle(med7_ner_data, "../data/clinical_notes/new_ner_word_dict.pkl")

# WORD2VEC EMBEDDING REPRESENTATION

med7_ner_data = pd.read_pickle("../data/clinical_notes/new_ner_word_dict.pkl")
w2vec = Word2Vec.load("../embeddings/word2vec.model")

data_types = [med7_ner_data]
data_names = ["new_ner"]
new_word2vec = {}

for data, names in zip(data_types, data_names):
    print("w2vec starting..")
    for k,v in data.items():

        patient_temp = []
        for i in v:
            if w2vec.wv.has_index_for(i[0]):
                patient_temp.append(w2vec.wv[i[0]])
            else:
                avg = []
                num = 0
                temp = []

                if len(i[0].split(" ")) > 1:
                    for each_word in i[0].split(" "):
                        if w2vec.wv.has_index_for(each_word):
                            temp = w2vec.wv[each_word]
                            avg.append(temp)
                            num += 1
                    if num == 0: continue
                    avg = np.asarray(avg)
                    t = np.asarray(list(map(mean, zip(*avg))))
                    patient_temp.append(t)
        if len(patient_temp) == 0: continue
        new_word2vec[k] = patient_temp

    print("word2vec length is {}", len(new_word2vec))
    pd.to_pickle(new_word2vec, "../data/clinical_notes/representations/" + names + "_word2vec_dict.pkl")

# FASTTEXT EMBEDDING REPRESENTATION

fasttext = FastText.load_fasttext_format("../embeddings/fasttext_skipgram.bin")

data_types = [med7_ner_data]
data_names = ["new_ner"]
new_fasttextvec = {}

for data, names in zip(data_types, data_names):
    print("fasttext starting..")        
    for k,v in data.items():
        patient_temp = []
        for i in v:
            if fasttext.wv.has_index_for(i[0]):
                patient_temp.append(fasttext.wv[i[0]])
        if len(patient_temp) == 0: continue
        new_fasttextvec[k] = patient_temp
    print("fasttext length is {}", len(new_fasttextvec))
    pd.to_pickle(new_fasttextvec, "../data/clinical_notes/representations/" + names + "_fasttext_dict.pkl")
 

# CONCAT EMBEDDING REPRESENTATION

data_types = [med7_ner_data]
data_names = ["new_ner"]
new_concatvec = {}

for data, names in zip(data_types, data_names):
    print("concat starting..")        
    for k,v in data.items():
        if k in new_word2vec and k in new_fasttextvec:
            patient_temp = []
            for i in v:
                w2vec_temp = []
                if w2vec.wv.has_index_for(i[0]):
                    w2vec_temp = w2vec.wv[i[0]]
                else:
                    avg = []
                    num = 0
                    temp = []

                    if len(i[0].split(" ")) > 1:
                        for each_word in i[0].split(" "):
                            if w2vec.wv.has_index_for(each_word):
                                temp = w2vec.wv[each_word]
                                avg.append(temp)
                                num += 1
                        if num == 0: 
                            w2vec_temp = [0] * 100
                        else:
                            avg = np.asarray(avg)
                            w2vec_temp = np.asarray(list(map(mean, zip(*avg))))
                    else:
                        w2vec_temp = [0] * 100
                
                if fasttext.wv.has_index_for(i[0]):
                    fasttemp = fasttext.wv[i[0]]
                else:
                    fasttemp = [0] * 100

                appended = np.append(fasttemp, w2vec_temp, 0)
                patient_temp.append(appended)
            if len(patient_temp) == 0: continue
            new_concatvec[k] = patient_temp

    print("concat length is {}", len(new_concatvec))
    pd.to_pickle(new_concatvec, "../data/clinical_notes/representations/" + names + "_concat_dict.pkl")

# DOC2VEC EMBEDDING REPRESENTATION

tag_data = [TaggedDocument(words=patient_clinical_notes[i], tags=[str(i)]) for i in patient_clinical_notes.keys()]

doc2vec_model = Doc2Vec(vector_size=100,
               alpha=0.025,
               min_alpha=0.00025,
               min_count=1,
               dm=1)

doc2vec_model.build_vocab(tag_data)
max_epochs = 25
for epoch in range(max_epochs):
        print(f"Iteration {epoch}")
        doc2vec_model.train(tag_data,
                   total_examples=model.corpus_count,
                   epochs=model.epochs)

        # Here I decrease the learning rate. 

        doc2vec_model.alpha -= 0.0002

        doc2vec_model.min_alpha = model.alpha
doc2vec_model.save("../embeddings/doc2vec.model")

new_doc2vec = {}
print("doc2vec starting..")        

for patient in patient_clinical_notes.keys():
    if count % 1000 == 0:
        print(count)
    new_doc2vec[patient] = list()
    new_doc2vec[patient].append(model.dv[str(patient)])

print("doc2vec length is {}", len(new_doc2vec))
pd.to_pickle(new_doc2vec, "../data/clinical_notes/representations/new_ner_doc2vec_dict.pkl")

intersect = set(new_word2vec.keys()) & set(new_fasttextvec.keys()) & set(new_concatvec.keys() & set(new_doc2vec.keys()))
new_word2vec_dict = {x: new_word2vec[x] for x in new_word2vec if x in intersect}
new_fasttextvec_dict = {x: new_fasttextvec[x] for x in new_fasttextvec if x in intersect}
new_concatvec_dict = {x: new_concatvec[x] for x in new_concatvec if x in intersect}
new_doc2vec_dict = {x: new_doc2vec[x] for x in new_doc2vec if x in intersect}
print (len(new_word2vec_dict), len(new_fasttextvec_dict), len(new_fasttextvec2_dict), len(new_concatvec_dict), len(new_doc2vec_dict))


pd.to_pickle(new_word2vec_dict, "../data/clinical_notes/representations/"+"new_ner"+"_word2vec_limited_dict.pkl")
pd.to_pickle(new_fasttextvec_dict, "../data/clinical_notes/representations/"+"new_ner"+"_fasttext_limited_dict.pkl")
pd.to_pickle(new_concatvec_dict, "../data/clinical_notes/representations/"+"new_ner"+"_combined_limited_dict.pkl")
pd.to_pickle(new_doc2vec_dict, "../data/clinical_notes/representations/"+"new_ner"+"_doc2vec_limited_dict.pkl")
